Moved mcfg io ui definitions
